Added box around Login, added nav bar, and moved how it works to homepage
Apologies for the multiple things happening in this commit. I updated
the nav bar and removed “how it works” and moved it to the home screen.
Reason being that when you are logged in you don’t necessarily need to
see how it works after you’ve already gone over it. Instead, when you
are on the home screen you can just scroll down to how it works. Also
put a border around the login button to make it look more engaging.