Added pictures to profile investments
Should have created a new branch for this, but got a little carried
away while I was working on the home formatting. Should be a lot more
clean going forward. Wanted to add a visual to the artists that you’ve
invested in. Not sure if I did this the right way, might want to double
check it.